Add undocumented KMS fields for Nitro Enclaves
AWS Nitro Enclaves make use of a set of currently undocumented request
and response fields for cryptographic attestation support in Decrypt,
GenerateDataKey, and GenerateRandom actions. This commit adds an interim
set of field and type definitions to support this information in KMS
requests until properly documented definitions are made available.